Fix the flaky dynamic params test thats been causing CI to fail (#568)
* Moving service_client.hpp to nav2_util

The ServiceClient class will be used Lifecycle helper functions that
will be put in nav2_util, so we need it there. It seemed more of utility
class than a Task anyhow.

* Add internal node helper functions.

Add some helper functions for generating node names. These will be used
to fix some potential problems with the ServiceClient and the upcoming
LifecycleServiceClient.

* Add node name filter to ServiceClient.

ServiceClient reuses the service name as the temporary node name.
However, service names frequently have / characters in them and
that is invalid for a node name. Also, it seems likely we'll have
multiple clients using the same service. This would result in a name
collision. This now adds a random suffix to the internal node name.

* Allow ServiceClient to use a passed in node.

ServiceClient can now use a supplied node instead of always creating
it's own. This allows us to create a higher level service client that
interacts with multiple services using just one generated node.

* Add a Lifecycle Service Client.

The unit tests will be handled at a higher level by a subsequent checkin

* Clean up and rename existing string utils module

I need to add another function to it, and it could stand some
restructuring before hand.

* Adding a split function

Adding a function to split a string at delimiters. The intent is to
use this later so tests can get ':' seperated string of lifecycle nodes
from and environment variable, and spin them all up.

* Add helper functions to facilitate bringup lifecycle nodes

* Add simple lifecycle bringup utility.

* Turn helper node into a lifecycle node.

* Add timeouts and retries to service calls.

Service calls still hang for some unknown reason. The only known
fix is to time it out and retry.

* Fixing coding standard violations in nav2_util library.

* Renaming one more function to comply with style guide.